Refactor PermisoDetalleChequeo to use Integer for idPermisoDetalle
- Changed idPermisoDetalle type from Long to Integer in PermisoDetalleChequeo and PermisoDetalleChequeoDTO classes.
- Updated mapping in PermisoTrabajoAlturasMapper to reflect the type change.

+++ back/src/main/java/com/seguridadlimite/models/permiso/domain/PermisoDetalleChequeo.java
@@ -20,7 +20,7 @@ public class PermisoDetalleChequeo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_permiso_detalle")
-    private Long idPermisoDetalle;
+    private Integer idPermisoDetalle;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_permiso")
